+++ fault/fault_test.go
@@ -10,6 +10,10 @@ const (
 	expectedFormat = "\n\nexpected:\n%s\n\nactual:\n%s\n\n"
 )
 
+// ------
+// User Error Tests
+// ------
+
 func Test_Error_WithSingleUserError(t *testing.T) {
 	code := "missing_first_name"
 	msg := "Please enter your first name."
@@ -145,10 +149,25 @@ func Test_ErrorMessages_WithMultipleUserErrors(t *testing.T) {
 	}
 }
 
+// ------
+// System Error Tests
+// ------
+
 func Test_Error_WithSingleSystemError(t *testing.T) {
 	f := System("a", "b", "c")
 
 	actual := f.Error()
+
+	expected := "a.b: c"
+	if actual != expected {
+		t.Errorf(expectedFormat, expected, actual)
+	}
+}
+
+func Test_String_WithSingleSystemError(t *testing.T) {
+	f := System("a", "b", "c")
+
+	actual := f.String()
 
 	expected := "a.b: c"
 	if actual != expected {
@@ -169,6 +188,19 @@ func Test_Error_WithLayersOfSystemErrorsAndOneNonSystemError(t *testing.T) {
 	}
 }
 
+func Test_String_WithLayersOfSystemErrorsAndOneNonSystemError(t *testing.T) {
+	f1 := errors.New("foo bar")
+	f2 := SystemWrap(f1, "d", "e", "f")
+	f3 := SystemWrap(f2, "g", "h", "i")
+
+	actual := f3.String()
+
+	expected := "g.h: i\n   d.e: f\n      foo bar"
+	if actual != expected {
+		t.Errorf(expectedFormat, expected, actual)
+	}
+}
+
 func Test_Error_WithLayersOfSystemErrors(t *testing.T) {
 	f1 := System("a", "b", "c")
 	f2 := SystemWrap(f1, "d", "e", "f")
@@ -179,5 +211,39 @@ func Test_Error_WithLayersOfSystemErrors(t *testing.T) {
 	expected := "g.h: i\n   d.e: f\n      a.b: c"
 	if actual != expected {
 		t.Errorf(expectedFormat, expected, actual)
+	}
+}
+
+func Test_FormatWithoutPlus_WithLayersOfSystemErrors_ReturnsSameAsError(t *testing.T) {
+	f1 := System("a", "b", "c")
+	f2 := SystemWrap(f1, "d", "e", "f")
+	f3 := SystemWrap(f2, "g", "h", "i")
+
+	expected := f3.Error()
+	notExpected := f3.StackTrace()
+
+	actual := fmt.Sprintf("%v", f3)
+	if actual != expected {
+		t.Errorf(expectedFormat, expected, actual)
+	}
+	if actual == notExpected {
+		t.Error("The 'v' formatter should not include a stack trace.")
+	}
+}
+
+func Test_FormatWithPlus_WithLayersOfSystemErrors_ReturnsSameAsStackTrace(t *testing.T) {
+	f1 := System("a", "b", "c")
+	f2 := SystemWrap(f1, "d", "e", "f")
+	f3 := SystemWrap(f2, "g", "h", "i")
+
+	expected := f3.StackTrace()
+	notExpected := f3.Error()
+
+	actual := fmt.Sprintf("%+v", f3)
+	if actual != expected {
+		t.Errorf(expectedFormat, expected, actual)
+	}
+	if actual == notExpected {
+		t.Error("The '+v' formatter should include a stack trace.")
 	}
 }